v1.3 of the Unofficial Doom Specs
--

The main reason for this release of the specs, 1.3, is of course the
explanation of the NODES structure. I've been delaying a little bit, because I
wanted to see if it would be feasible to include a good algorithm herein.
Also, I wanted to wait and see if someone could actually implement "node
theory" in a level editor, thereby verifying it.

Now the theory HAS been verified. However, the actual implementation is still
being worked on (debugged) as I'm writing this. Also, I don't want to steal
anyone's hard work outright. This means that there is NOT a node creation
algorithm here, but I do outline how one can be done. I have tried to come up
with one on my own, but it is too difficult for me, especially with all the
other things I'm simultaneously doing.

Where you WILL find pseudo-code is in the BLOCKMAP section. I borrowed an
excellent idea from a contributor, and code based on the algorithm given here
should be very fast. Even huge levels should recalculate in seconds.

Another new section completely explains the REJECT resource.

This entire document has been re-formatted, and there have been several other
additions, and hopefully the last of the typos has been rooted out. I consider
these specs to be at least 95% complete. There are only minor gaps in the
information now. If the promised "official specifications" were released
today, I expect this would compare favorably with them (although I know
exactly what parts of it I would look to first).

I've been notified of something very disappointing, and after a couple weeks
of trying there seems to be no way around it. The pictures that are used for
sprites (things like barrels, demons, and the player's pistol) all have to be
listed together in one .WAD file. This means that they don't work from pwad
files. The same thing goes for the floor pictures. Luckily, the walls are done
in a more flexible way, so they work in pwads. All this is explained in
chapter [5].